requirements.txt
streamlit
langchain
langchain-google-genai  # New library for Gemini LLM/Agents
google-genai            # New core Google GenAI SDK
faiss-cpu
python-dotenv